api: Add schema for inspection rules API (requests)
Much of the validation already exists here. We're just shuffling it
about. We do end up with some validation, particularly with regards to
the code in ironic.common.inspection_rules.validation, but we can clean
that up in future changes once we've decided whether we're okay with
error message content changing or not (hopefully yes).

Change-Id: Id97b6d524913188cec557e0df64ebc1a2fc3eccd
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>