Cleanup tests for eventlet migration
In preparing to remove eventlet, we discoverd eventlet was hiding
three distinct issues in our tests.

1) Eventlet was hiding some issues, like... some tests explicitly
   trying to stop a fake service.
2) Our tests at time crossed the boundry of a thread. While not
   awful, and provides us a higher level of assurance.
3) Some of the bad patterning sort of amplified in some of the code
   which we'll eventually need to sort out.

In any event, The cleanest path was to make a test centric start
method which does the basic needful for the tests, *and* which also
explicitly launches Synchronous Executors as opposed to threads.

This pattern can be leveraged to start to sub divide the actual
service launches at some point, while also keeping explicit
synchronous executor behavior which will be absolutely required
for use of threads in the post-eventlet world.

Change-Id: Ib75406a3e386b78197ea78cce291c3bd295d20cf
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>