ci: stabilize ironic-standalone-redfish
While looking at issue reports, I noticed we are likely stressing
the ramdisk too much and running it out of space.

Anyhow, one of our CI jobs fails more than others, and it needs to
be swapped around so the image is not downloaded, then converted in
the ramdisk. It is the only job which does it, and we should keep that
behavior, but we need to get CI in a happier place first.

Related-bug: 2116135
Change-Id: I77c30c370cf5288703663e495ab9e60f3e8a7b2e
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>