ci: mark failing jobs non-voting pending fixes/removal
The snmp job has started failing because we don't have
an event reconcilation loop for asyncio in the main service.
It seems somehow the asyncio version of pysnmp is slipping into
the CI job which ultimately is breaking us at this point, although
I see no actual state of it in the logs. Its just weird.

In any event, it is a known issue.

The other issue is the grenade jobs are failing on neutron upgrades.
This appears to be due to legacy names being used in the CI job
configuration.

We should fix those, but we've got an open bug for that now:
https://bugs.launchpad.net/ironic/+bug/2118780

Change-Id: I1fbe4b0c519b5911db6f92e2963df99a882fa317
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>