Remove direct mapping from API -> DB
In order to complete the eventlet migration, we need to
be able to use the indirection_api surface to be able to
disjoint the API surface from the message bus, so we're
not fighting between processes for database locks.

Previously, the calls were directly routed directly
to the database layer to streamline performance.
This attachment allows the calls to run through
the indirection layer to the object, and return
the data to the caller.

As a result, API RPC versions are incremented.

Change-Id: I7358ac2a70198c78a0a9ba48511fc1289c64294f
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>