Revert "ci: stabilize ironic-standalone-redfish"
This reverts commit 03fd6f992454fe2d609efcebc81bf0b627ea2abd
which was put in place to temporarily lock the data flow for
IPA agents while we worked to reduce the overall ramdisk size.

This is done with https://review.opendev.org/c/openstack/ironic-python-agent-builder/+/954524
and https://review.opendev.org/c/openstack/ironic-python-agent-builder/+/952756
which should reduce the overall runtime memory footprint.

Change-Id: I2ac10de2616a6579983c7f3199d285346c84adbf
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>