ci: move metal3 job to larger memory node
The meteal3 job has traditionally operated as a very tight memory
condition for quite some time. Specifically running minikube, ironic,
and two VMs in 8 GB of memory also influenced the job design.

However, with the removal of eventlet, this memory footprint and
process model swells a little bit creating conditions where the
job is essentially guarenteed to fail. Further discussion with
Riccardo, one of the other Ironic/Metal3 contributros yielded
that they were already thinking of increasing the size of the
VM because they had been encountering memory issues.

Given that this job already runs lean, albeit with no swap
which is intentional to prevent disk IO generated noisey neighbor
conditions, the only real choice is to just increase the VM size
to the next realistic size in OpenDev CI.

Change-Id: I87f9c94e6585347d8a35a1d04dd7d101a9e68261
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>