ci: temporary metal3 integration job disable
In order to merge eventlet removal patches and fixes related to
metal3, we need to merge four separate changes in series to get
the metal3 CI job back into a happy state. This is because:
* Eventlet removal changes the process model to sub-processes
* Metal3 Integration uses sqlite, which we've learned in the past
  can have locking issues between processes.
* The fix requires removal of direct database calls from the
  API surface, and instead for calls to be routed through the RPC
  layer.

Once the changes are merged together, the metal3 job has been shown
to work in other test runs, so we have high confidence overall, we
just unfortunately need give ourselves the window where the job is
not passing to merge sequence of changes which will be stacked
after this change.

Change-Id: I3e6cb1c25b04ff965fa40ff6dbac9bd1bb53c44b
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>