Launch vnc proxy with no_fork
The VNCProxy does some things which must be done in the parent
process, i.e. signal handling. The only compromise to keep it
in the parent process, is to utilize no_fork.

This is a compromise as the new process model works nicely with
systemd, so only tries no_fork if VNC is enabled with the single
process mode.

Change-Id: I45aa9f4ffad946c59ca3f551c971b979fbc64efd
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>