Memoize calls to bcrypt.checkpw
This intentionally high CPU overhead function is called for every API
and JSON-RPC request when Basic HTTP authentication is enabled. With the
recent indirection enablement this is causing a performance regression
for Metal3 due to the extra JSON-RPC calls. This change would improve
the performance of all branches of Metal3 if backported.

(JayF) In 2024.2, this code actually lived in ironic-lib. Instead of
patching it there and requiring a release update, I thought it might be
preferable to fix this directly in Ironic.

Change-Id: I2740035d2882aacddca9c541362d6e533140650f
Closes-Bug: #2121105
Signed-off-by: Steve Baker <sbaker@redhat.com>
Signed-off-by: Jay Faulkner <jay@jvf.cc>
(cherry picked from commit 2a62718887f0bd38fe61671d221121268008cae1)