Launch API in the same process as conductor for singleprocess Ironic
Indirection proved to impose very high performance costs. Even though
we've identified and fixed a few major contributors to the regression,
using RPC for a process to access itself has always been a stopgap
measure rather than a proper architecture.

This change moves the API from a separate service under the Launcher
to a separate thread in the conductor, similarly to how the JSON RPC
server is started. This paves the way to remove the local RPC entirely
in the next patch.

Change-Id: I0f3d854336bcc7ea1062f9a995e6d8979cb0cc22
Signed-off-by: Dmitry Tantsur <dtantsur@protonmail.com>