Improve logging when console unit doesn't start
Log the unit status and journal log when systemctl start fails for a
console container.

Also before the console is stopped, call journalctl for that unit. If
debug is enabled then the journal output will be logged.

Additionally, after a container is started, attempts are made to open a
socket to the VNC port and read some data. There is a delay between the
container starting and x11vnc actually listening and this race can be
triggered in automated tests, so this delays changing the console
enabled state until it is *really* ready.

Change-Id: I2c4867b6773f4f4eaa8b98e50a63881f0f4d08b0
Signed-off-by: Steve Baker <sbaker@redhat.com>