Set columns in bios_settings table to bigint
Currently the columns upper_bound and lower_bound are set to a signed
int, causing issues for systems reporting larger values. The values can
be up to 32-bit unsigned as per redfish standard.

This patch will change the data type to BigInteger, which will allow the
larger values to be stored.

Change-Id: Ife9742784c8d59fd5c6b89757e215938ec0a595d
Closes-Bug: #2095486
Signed-off-by: Freerk-Ole Zakfeld <fzakfeld@scaleuptech.com>
Signed-off-by: ricolin <rlin@vexxhost.com>