fix: glance image member lookup resulted in an empty list always
The code did not initialize a client so it resulted in an exception
always occurring. It also used the image attribute on the client
attribute of the service but the client attribute of the service is
already the image attribute. Create a wrapper method to use the API
correctly and prevent similar issues.

Closes-Bug: #2099276
Closes-Bug: #2132246
Change-Id: Ib803c066ca28d1c05a345b7a982a0daabbd7d52e
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>
(cherry picked from commit 7ce4fdfb100380eddbbddd548e5957c34597d1f0)