fix: ensure that the bootloader path exists
Ensure that the path where we are copying the bootloader into exists
before we copy the file in. Unfortunately the existing code wouldn't
work in the intended case because os.path.split("snponly.efi") returns
("", "snponly.efi") so we would never pass the first check so we would
not create the path. This keeps the same behavior of allowing a nested
path structure to not change the behavior so that this can be
backportable while fixing the issue and switching the pathlib to provide
a safer interface going forward.

Conflict: ironic/common/pxe_utils.py - Python 3.9 did not support the
type annotation so I've removed it

Change-Id: Iaf756f634832310431020abd758b59c749aecb21
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>
(cherry picked from commit 7a0a9ef784a45451dac2562507482c84d033b39e)