fix redfish inspect system product name
The intention of this code is to read the system product name which is
stored in the model field per the Redfish spec and not in the name field
which will always store the name of the object we are working with. This
results in the value always being 'System'.

Conflict:
  include a change to the mocked system object that happened in a
  feature change patch that cannot be backported

Closes-Bug: 2136233
Change-Id: I375fbe27253d7965e458be7b147d5b72cffa4e89
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>
(cherry picked from commit ade41cae4760ef371889f6d2701f54b6a308b807)